% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bfcluster}
\alias{bfcluster}
\title{Birkhoff-Fan clustering}
\usage{
bfcluster(S, nclust = as.integer(c()), maxiter = 100L, tolerance = 0.01,
  admm_penalty = 100, verbose = 0L)
}
\arguments{
\item{S}{input pairwise similarity matrix
(assumed to be symmetric)}

\item{nclust}{a vector of the number of clusters}

\item{maxiter}{max number of iterations for each solution}

\item{tolerance}{convergence threshold}

\item{admm_penalty}{penalty parameter of ADMM algorithm}

\item{verbose}{level of verbosity}
}
\value{
An S3 object of class \code{bfcluster} which is a list with 
        the following components:
 \item{nclust}{a vector containing the number of clusters of each estimate}
 \item{clustmat}{a list containing the normalized clustering matrix 
                 estimates}
 \item{maxval}{a vector of optimal objective function values 
               for each value of the number of clusters}
 \item{niter}{a vector containing the number of ADMM iterations for each 
              estimate}
}
\description{
This function computes a solution sequence of the Birkhoff-Fan clustering. 
It takes a symmetric matrix \code{S} as input and returns a object 
containing a list of normalized clustering matrices estimated by 
Birkhoff-Fan clustering over a sequence of the number of clusters.
}

