% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering.bfcluster}
\alias{clustering.bfcluster}
\title{Extract a normalized clustering matrix estimate}
\usage{
\method{clustering}{bfcluster}(object, k, ...)
}
\arguments{
\item{object}{bfcluster object}

\item{k}{number of clusters}

\item{...}{other arguments}
}
\description{
Returns an estimated normalized clustering matrix
}

