% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Extract a clustering matrix estimate}
\usage{
clustering(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\description{
\code{clustering} is a generic function for extracting an estimated clustering matrix from a clustering object
}

