% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_miscls.R
\name{count_miscls}
\alias{count_miscls}
\title{Count the number of misclassified objects}
\usage{
count_miscls(label_est, label_true)
}
\arguments{
\item{label_est}{a vector of estimated labels}

\item{label_true}{a vector of true labels}
}
\value{
The number of misclassified objects
}
\description{
Count the number of misclassified objects comparing to provided true labels
}

