% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_adjmat.R
\name{gen_adjmat}
\alias{gen_adjmat}
\title{Generate a network edge probability matrix and a random adjacency matrix}
\usage{
gen_adjmat(n, k, sizes = NULL, B = NULL, p_min, p_max, q_min, q_max,
  type = c("plantedsbm", "sbm", "assortative"), self_loop = FALSE,
  perm = FALSE)
}
\arguments{
\item{n}{the number of nodes}

\item{k}{the number of communities}

\item{sizes}{a vector of community sizes}

\item{B}{a matrix of community-wise edge probabilities.
Ff supplied, it overrides \code{p_min}, \code{p_max}, 
\code{q_min}, \code{q_max}, and \code{type} arguments.}

\item{p_min}{the minimum within-community edge probability}

\item{p_max}{the maximum within-community edge probability. 
It has no effect if argument \code{type} is "plantedsbm".}

\item{q_min}{the minimum between-community edge probability. 
It has no effect if argument \code{type} is "plantedsbm".}

\item{q_max}{the maximum between-community edge probability}

\item{type}{type of network model. It has to be "plantedsbm", "sbm", 
or "assortative". It has no effect if argument \code{B} 
is supplied.}

\item{self_loop}{a logical indicating whether to allow self loops 
in the network}

\item{perm}{a logical indicating whether to randomly permute the generated
edge probability matrix, random adjacency matrix, membership
matrix, and label}
}
\value{
A list with the following components:
  \item{adjmat}{a random adjacency matrix if \code{sample} is TRUE}
  \item{probmat}{a network edge probability matrix}
  \item{membermat}{the membership matrix according to node labels}
  \item{label}{the label vector of nodes}
}
\description{
Generate a network edge probability matrix from a specified network model with community structure and a random adjacency matrix
}

