% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getlab.bfcluster}
\alias{getlab.bfcluster}
\title{Compute label estimates}
\usage{
\method{getlab}{bfcluster}(object, k, normalize = TRUE, iter_max = 1000,
  rep = 100, ...)
}
\arguments{
\item{object}{bfcluster object}

\item{k}{number of clusters}

\item{normalize}{a logical indicating whether to normalize the 
rows of clustering matrix to have unit norms}

\item{iter_max}{the maximum number of iterations allowed for k-means
algorithm}

\item{rep}{number of k-means replicates}

\item{...}{other arguments}
}
\description{
Produces label estimates using k-means algorithm based on an estimated normalized clustering matrix
}

