% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getlab}
\alias{getlab}
\title{Compute label estimates}
\usage{
getlab(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\description{
\code{getlab} is a generic function for extracting label estimates from a clustering object
}

