% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute.bfcluster}
\alias{permute.bfcluster}
\title{Permute a normalized clustering matrix estimate}
\usage{
\method{permute}{bfcluster}(object, k, ...)
}
\arguments{
\item{object}{bfcluster object}

\item{k}{number of clusters}

\item{...}{optional arguments to \code{getlab} method}
}
\value{
A list with the following arguments:
  \item{perm_mat}{permuted normalized clustering matrix estimate}
  \item{perm_label}{a vector of permuted label estimates}
  \item{perm_id}{a vector of indices based on which to permute the 
                 normalized clustering matrix and the label vector}
}
\description{
Produces a permuted normalized clustering matrix according to label estimates computed from it
}

