% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Permute a clustering matrix estimate}
\usage{
permute(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{other arguments}
}
\description{
\code{permute} is a generic function for permuting an estimated clustering matrix according to label estimates computed from it
}

