% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bfcluster}
\alias{plot.bfcluster}
\title{Plot bfcluster object}
\usage{
\method{plot}{bfcluster}(object, k = NULL, prob_thres = NULL,
  perm = FALSE, ...)
}
\arguments{
\item{object}{bfcluster object}

\item{k}{the number of clusters. If not provided, 
it is set to be the first value of 
the number of clusters in the object}

\item{prob_thres}{the probability at which to threshold the large entries
of the clustering matrix for better visualization}

\item{perm}{a logical indicating whether to permute the 
estimated normalized clustering matrix 
according to label estimates 
computed from it before plotting}

\item{...}{further arguments passed to \code{permute} method}
}
\description{
Plot an estimated normalized clustering matrix for a specified number of clusters
}

