% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.matrix}
\alias{plot.matrix}
\title{Plot matrix object}
\usage{
\method{plot}{matrix}(S, ...)
}
\arguments{
\item{S}{matrix object}

\item{...}{other arguments passed to or from other methods}
}
\description{
This function plots a matrix
}

